Creating ways to deal with exceptions

+++ src/main/java/com/apirestful/crud/controller/UsersController.java
@@ -4,10 +4,15 @@ package com.apirestful.crud.controller;
 import com.apirestful.crud.model.User;
 import com.apirestful.crud.repository.UserRepository;
 import com.apirestful.crud.service.UserService;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.swing.text.html.Option;
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/users")
@@ -18,29 +23,54 @@ public class UsersController {
     @Autowired
     private UserService userService;
 
+
+
     @GetMapping()
     public List<User> getUsers(){
        return userService.getAllUsers();
     };
 
     @GetMapping("/{id}")
-    public User getUserById(@PathVariable int id){
-        return userService.getUserById(id);
+    public ResponseEntity<Optional<User>> getUserById(@Valid @PathVariable long id){
+        Optional<User> user = Optional.ofNullable(userService.getUserById(id));
+        if(user.isPresent()){
+            return ResponseEntity.status(HttpStatus.OK).body(user);
+        };
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                .body(null);
     };
 
     @PostMapping()
-    public User createUser(@RequestBody User user){
-        return userService.createUser(user);
+    public ResponseEntity<Optional<User>> createUser(@Valid @RequestBody User user){
+        Optional<User> newUser = Optional.ofNullable(userService.createUser(user));
+
+        if(newUser.isPresent()){
+            return ResponseEntity.status(HttpStatus.CREATED).header("Location", "/users/" + newUser.get())
+                    .body(newUser);
+        };
+
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
     };
 
     @PutMapping("/{id}")
-    public User updateUser(@PathVariable long id, @RequestBody User user){
-        return userService.updateUser(user, id);
+    public ResponseEntity<Optional<User>> updateUser(@PathVariable long id,@Valid @RequestBody User user){
+        Optional<User> newUser = Optional.ofNullable(userService.updateUser(user,id));
 
+        if(newUser.isPresent()){
+            return ResponseEntity.status(HttpStatus.OK).body(newUser);
+        };
+
+        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
     };
 
     @DeleteMapping("/{id}")
     public void deleteUser(@PathVariable long id){
-        userService.deleteUserById(id);
+        User user = userService.getUserById(id);
+
+        if(user != null){
+            userService.deleteUserById(id);
+            ResponseEntity.status(HttpStatus.OK);
+        } else
+            ResponseEntity.status(HttpStatus.NOT_FOUND);
     };
 }
